requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
SQLAlchemy[asyncio]>=2.0.29
aiosqlite>=0.19.0
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=1.10.14
pydantic-settings>=2.1.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
